+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### v1.6.0 (in development)
 
+- Added a way to pass function to props in definition.
 - Added a way to pass function to `optional` prop at elem definition.
 - Added a way to disable an element by calling it as `false` or `null`.
 - Added docs for the external `bemto-button` component.